+++ sphericalenvmapping/sphericalenvmapping.cpp
@@ -210,7 +210,7 @@ public:
 
 		VkDescriptorPoolCreateInfo descriptorPoolInfo =
 			vks::initializers::descriptorPoolCreateInfo(
-				poolSizes.size(),
+				static_cast<uint32_t>(poolSizes.size()),
 				poolSizes.data(),
 				2);
 
@@ -236,7 +236,7 @@ public:
 		VkDescriptorSetLayoutCreateInfo descriptorLayout =
 			vks::initializers::descriptorSetLayoutCreateInfo(
 				setLayoutBindings.data(),
-				setLayoutBindings.size());
+				static_cast<uint32_t>(setLayoutBindings.size()));
 
 		VK_CHECK_RESULT(vkCreateDescriptorSetLayout(device, &descriptorLayout, nullptr, &descriptorSetLayout));
 
@@ -323,7 +323,7 @@ public:
 		VkPipelineDynamicStateCreateInfo dynamicState =
 			vks::initializers::pipelineDynamicStateCreateInfo(
 				dynamicStateEnables.data(),
-				dynamicStateEnables.size(),
+				static_cast<uint32_t>(dynamicStateEnables.size()),
 				0);
 
 		// Spherical environment rendering pipeline
@@ -456,7 +456,7 @@ public:
 		{
 			uboVS.texIndex = textures.matCapArray.layerCount-1;
 		}
-		if (uboVS.texIndex >= textures.matCapArray.layerCount)
+		if (uboVS.texIndex >= (int32_t)textures.matCapArray.layerCount)
 		{
 			uboVS.texIndex = 0;
 		}
